Ensure L337 can only give shields to L337
This doesn't 100% fix the edge case - if someone plays a L337 while having one on the board, and choosing to sacrifice the new L337, you can give the shield to the L337 on the board already. This has proven very tricky to plan around, but this improvement fixes most issues.