Bug Fixes Round 3 (#324)
* Jabba's Rancor and Death Trooper are not optional

* Fix Arquitens Assault Cruiser's resource entering play readied

* Ensure L337 can only give shields to L337

This doesn't 100% fix the edge case - if someone plays a L337 while having one on the board, and choosing to sacrifice the new L337, you can give the shield to the L337 on the board already. This has proven very tricky to plan around, but this improvement fixes most issues.

* Fix Lurking TIE Phantom

There was previously an edge case where the Lurking TIE Phantom could be damaged by enemy damage triggered on the owner's turn, as well as not be damaged by friendly damage triggered on the enemy turn.

* Fix DJ

The old implementation only worked for the current turn. This will guarantee the Resource has a reference to the unique ID of the DJ that stole it.

---------

Co-authored-by: Dan Bastin <dan.bastin@gmail.com>